Extract AbstractDocumentServiceRegistry
CmaCaseServiceRegistry and AaibReportServiceRegistry do the same thing
with variation for just three dependencies -- repository, builder and
observers.
