+++ app/services/cma_case_service_registry.rb
@@ -1,92 +1,18 @@
+require "abstract_document_service_registry"
 require "specialist_publisher_wiring"
-require "list_documents_service"
-require "show_document_service"
-require "new_document_service"
-require "preview_document_service"
-require "create_document_service"
-require "update_document_service"
-require "publish_document_service"
-require "withdraw_document_service"
 require "cma_case_observers_registry"
 
-class CmaCaseServiceRegistry
-  def list
-    ListDocumentsService.new(
-      cma_case_repository,
-    )
-  end
-
-  def show(document_id)
-    ShowDocumentService.new(
-      cma_case_repository,
-      document_id,
-    )
-  end
-
-  def new
-    NewDocumentService.new(
-      cma_case_builder,
-    )
-  end
-
-  def preview(document_id, attributes)
-    PreviewDocumentService.new(
-      cma_case_repository,
-      cma_case_builder,
-      document_renderer,
-      document_id,
-      attributes,
-    )
-  end
-
-  def create(attributes)
-    CreateDocumentService.new(
-      cma_case_builder,
-      cma_case_repository,
-      observers.creation,
-      attributes,
-    )
-  end
-
-  def update(document_id, attributes)
-    UpdateDocumentService.new(
-      repo: cma_case_repository,
-      listeners: observers.update,
-      document_id: document_id,
-      attributes: attributes,
-    )
-  end
-
-  def publish(document_id)
-    PublishDocumentService.new(
-      cma_case_repository,
-      observers.publication,
-      document_id,
-    )
-  end
-
-  def withdraw(document_id)
-    WithdrawDocumentService.new(
-      cma_case_repository,
-      observers.withdrawal,
-      document_id,
-    )
-  end
-
+class CmaCaseServiceRegistry < AbstractDocumentServiceRegistry
 private
-  def observers
-    @observers ||= CmaCaseObserversRegistry.new
-  end
-
-  def cma_case_repository
+  def document_repository
     SpecialistPublisherWiring.get(:cma_case_repository)
   end
 
-  def cma_case_builder
+  def document_builder
     SpecialistPublisherWiring.get(:cma_case_builder)
   end
 
-  def document_renderer
-    SpecialistPublisherWiring.get(:specialist_document_renderer)
+  def observers
+    @observers ||= CmaCaseObserversRegistry.new
   end
 end
